fix: audit-surfaced game-logic and constant bugs
Bugs the normalized metric was masking (all verified via objdiff):
- SongParser::TrackAllowsOverlappingNotes: removed spurious kTrackVocals case
  (target allows only kTrackKeys/kTrackRealKeys); 99.62 -> 100%. Real runtime
  bug: vocals were incorrectly told they could have overlapping notes.
- GameConfig::GetFxSwitchPosition: dual float->int recast so both operands of
  the sign-mask materialize their own stack temp; 97.8 -> 99.6%.
- SetlistSort::BuildSetlistTree: removed bogus (SongSortNode*) sibling-class
  cast that selected the wrong NewShortcutNode overload (vtable slot 0xc4 vs
  0xf0); -> 100%. SongSort unit now 100%.
- SongStatusMgr/WaitingUserGate/Tour: corrected MILO_ASSERT/HANDLE_CHECK line
  numbers to match target.

+++ src/system/beatmatch/SongParser.cpp
@@ -1585,7 +1585,7 @@ bool SongParser::CheckForceHopoMarker(int tick, int pitch, bool b) {
 bool SongParser::CheckDrumMapMarker(int i, int j, bool b) {
     if (!mDrumStyleGems)
         return false;
-    else if ((int)(int)(int)108U <= j & j < 113U) {
+    else if (108 <= j & j < 113U) {
         mSink->DrumMapLane(mTrack, i, j - 108, b);
         return true;
     } else
